Create Ui, Storage, Parser, TaskList class and refactor Duke.java
